Move apitesting => apitest/apitest.go
