+++ core/internal/apitest/apitest.go
@@ -1,21 +1,22 @@
-package api
+package apitest
 
 import (
 	"net/url"
 
 	"github.com/hashicorp/go-retryablehttp"
+	"github.com/wandb/wandb/core/internal/api"
 )
 
 // Returns a fake [Client] for testing.
 func FakeClient(
 	baseURLString string,
 	retryableHTTP *retryablehttp.Client,
-) *Client {
+) *api.Client {
 	baseURL, err := url.Parse(baseURLString)
 	if err != nil {
 		panic(err)
 	}
 
-	backend := New(baseURL)
+	backend := api.New(baseURL)
 	return backend.NewClient(retryableHTTP)
 }
