+++ core/pkg/filestream/filestream_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/wandb/wandb/core/internal/api"
+	"github.com/wandb/wandb/core/internal/apitest"
 	"github.com/wandb/wandb/core/internal/clients"
 
 	"github.com/wandb/wandb/core/pkg/observability"
@@ -147,7 +147,7 @@ func NewFilestreamTest(tName string, fn func(fs *filestream.FileStream)) *filest
 		filestream.WithPath(fstreamPath),
 		filestream.WithSettings(tserver.settings),
 		filestream.WithLogger(tserver.logger),
-		filestream.WithAPIClient(api.FakeClient(
+		filestream.WithAPIClient(apitest.FakeClient(
 			tserver.hserver.URL,
 			clients.NewRetryClient(
 				clients.WithRetryClientHttpAuthTransport(tserver.settings.GetApiKey().GetValue()),
