+++ core/pkg/filestream/loop_transmit_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/hashicorp/go-retryablehttp"
-	"github.com/wandb/wandb/core/internal/api"
+	"github.com/wandb/wandb/core/internal/apitest"
 	"github.com/wandb/wandb/core/internal/clienttest"
 	"github.com/wandb/wandb/core/pkg/observability"
 )
@@ -63,7 +63,7 @@ func testSendAndReceive(t *testing.T, chunks []processedChunk, fsd FsTransmitDat
 
 	fs := NewFileStream(
 		WithLogger(fsTest.logger),
-		WithAPIClient(api.FakeClient("", fsTest.client)),
+		WithAPIClient(apitest.FakeClient("", fsTest.client)),
 	)
 	for _, d := range chunks {
 		fs.transmitChan <- d
